+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file. The format
 is based on [Keep a Changelog](https://keepachangelog.com).
 
-## [Unreleased]
+## [0.19.1] - 2023-05-01
 
 ### Added
 
@@ -966,7 +966,8 @@ is based on [Keep a Changelog](https://keepachangelog.com).
 - Setting the log level to `quiet` now properly suppresses any log output.
 - Configuring colored terminal output should now print colored output.
 
-[Unreleased]: https://github.com/actor-framework/actor-framework/compare/0.19.0...master
+[Unreleased]: https://github.com/actor-framework/actor-framework/compare/0.19.1...master
+[0.19.1]: https://github.com/actor-framework/actor-framework/releases/0.19.1
 [0.19.0]: https://github.com/actor-framework/actor-framework/releases/0.19.0
 [0.19.0-rc.1]: https://github.com/actor-framework/actor-framework/releases/0.19.0-rc.1
 [0.18.7]: https://github.com/actor-framework/actor-framework/releases/0.18.7